requirements.txt
fastapi
uvicorn
python-dotenv
numpy
sentence-transformers
cohere
